More eval optimizations

+++ src/search-eval/search.cpp
@@ -249,7 +249,7 @@ Move search(Board& board, int max_depth, int& best_score) {
         }
 
         std::cout << "info depth " << depth << " seldepth " << seldepth << " score " << best_score << " nodes " << nodes << " nps "
-                  << (!duration.count() ? 0 : (static_cast<double>(nodes) / (static_cast<double>(duration.count()) / 1000))) << " pv ";
+                  << (!duration.count() ? 0 : static_cast<int>(static_cast<double>(nodes) / (static_cast<double>(duration.count()) / 1000))) << " pv ";
         for (uint16_t i = 0; i < pv_table[0].cur_move; i++) {
             std::cout << moveToStr(pv_table[0].moves[i]) << ' ';
         }
